refactor: move HttpInterfaceConfig configuration parser entity to own package
