+++ src/main/java/org/hyperagents/yggdrasil/utils/HttpInterfaceConfig.java
@@ -1,0 +1,13 @@
+package org.hyperagents.yggdrasil.utils;
+
+import java.util.Optional;
+
+public interface HttpInterfaceConfig {
+  String getHost();
+
+  int getPort();
+
+  String getBaseUri();
+
+  Optional<String> getWebSubHubUri();
+}
